+++ src/mako/validator/rules/BooleanFalse.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\validator\rules;
+
+use function sprintf;
+
+/**
+ * Boolean false rule.
+ */
+class BooleanFalse extends Rule implements RuleInterface
+{
+	/**
+	 * {@inheritDoc}
+	 */
+	public function validate($value, array $input): bool
+	{
+		return $value === false;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function getErrorMessage(string $field): string
+	{
+		return sprintf('The %1$s field must contain the boolean value FALSE.', $field);
+	}
+}